Update README.md
Ensure that the AdminReceiver class is properly implemented and registered in the manifest file.

Add checks to verify if the accelerometer sensor is available before registering it. You can use the getDefaultSensor() method to check if it returns null before registering.

Consider displaying a Toast message when the screen is locked to indicate to the user that the action has been successfully performed.